Module Metrics_mirage.Influx

Influx reporter

Parameters

CLOCK : Mirage_clock.MCLOCK

STACK : Mirage_stack.V4

Signature

val vmname : string -> Metrics.field

vmname name creates a tag with the virtual machine name.

val create : STACK.t -> ?⁠interval:int -> ?⁠hostname:string -> STACK.TCPV4.ipaddr -> ?⁠port:int -> unit -> (Metrics.reporter, unit) Stdlib.result Lwt.t

create mclock stack ~interval ~hostname ip ~port () is reporter, which sends measurements (prefixed by vmname hosttname if provided), to ip:port (defaults to 8094). If interval is provided, measurements from each source are at most reported once during interval milliseconds. Fails if the initial connection attempt fails. Reconnects TCP session on failure.